Meeting

+++ src/main/java/com/cmpn306/database/WebGraphNode.java
@@ -3,11 +3,12 @@ package com.cmpn306.database;
 import java.util.Hashtable;
 
 public class WebGraphNode {
-    String docUrl;
-    float pageRank;
-    Hashtable<String,WebGraphNode> outGoingUrls;
-    public WebGraphNode(String docUrl,float pageRank){
-        this.docUrl = docUrl;
+    String                          docUrl;
+    double                          pageRank;
+    Hashtable<String, WebGraphNode> outGoingUrls;
+
+    public WebGraphNode(String docUrl, double pageRank) {
+        this.docUrl   = docUrl;
         this.pageRank = pageRank;
     }
